types: widen the compression byte before to_string
Third site of the same gcc refusal this branch has hit: a narrow unsigned
passed where both an int and an unsigned int overload are viable makes the
promotion pick int, and -Wsign-promo rejects the choice. physical_value and
vector_operations were the first two; each only became visible after the
previous one was fixed, because the build stops at the first error.

The neighbouring dispatch in column_segment.cpp already casts before
to_string. This is the same message built one file over, without the cast.

+++ components/table/storage/data_pointer.cpp
@@ -52,7 +52,8 @@ namespace components::table::storage {
             // The byte is DISK-FED and semantically unknown to this reader; the CRC cannot
             // catch it (a new writer wrote it deliberately). Refuse the load loudly instead
             // of scanning the stream as raw fixed-size bytes.
-            auto message = "column pointer names compression byte " + std::to_string(compression_byte) +
+            auto message = "column pointer names compression byte " +
+                           std::to_string(static_cast<unsigned>(compression_byte)) +
                            ", which this reader does not implement";
             reader.latch_corruption(message.c_str());
             return result;
